pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "permafrost"
version = "1.0.0"
description = "Permafrost — Open-source AI brain framework with multi-provider, multi-channel support."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "dead1786"},
]
keywords = ["ai", "chatbot", "brain", "multi-agent", "telegram", "discord"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "requests>=2.31.0",
]

[project.optional-dependencies]
claude = ["anthropic>=0.40.0"]
openai = ["openai>=1.50.0"]
gemini = ["google-generativeai>=0.5.0"]
telegram = []
discord = []
console = ["streamlit>=1.30.0"]
all = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "google-generativeai>=0.5.0",
    "streamlit>=1.30.0",
]

[project.scripts]
permafrost = "permafrost.launcher:main"

[project.urls]
Homepage = "https://github.com/dead1786/permafrost"
Issues = "https://github.com/dead1786/permafrost/issues"

[tool.setuptools.packages.find]
include = ["core*", "channels*", "smart*", "console*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
